#2724: cleanup
